Projects/AI_CV_Critic/pyproject.toml
[project]
name = "ai-cv-critic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-groq>=0.3.2",
    "openai>=1.82.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
